backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "celery>=5.6.0",
    "fastapi>=0.127.0",
    "faster-whisper>=1.2.1",
    "passlib>=1.7.4",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pydub>=0.25.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.40.0",
]
